REFACTOR: Decomposed mappers into separate files and put into mappers directory
