+++ actions/playlist/create-playlist.ts
@@ -1,6 +1,6 @@
 import { createServerSupabaseClient } from "@/utils/supabase/server";
 import type { Playlist } from "@/types/playlist";
-import { mapPlaylistRow } from "@/lib/mappers";
+import { mapPlaylistRow } from "@/lib/mappers/playlist";
 import { CreatePlaylistSchema } from "@/schemas/playlists/create-playlist.schema";
 
 /**
