+++ app/albums/[id]/components/AlbumDetailContent.tsx
@@ -8,7 +8,7 @@ import { toast } from "sonner";
 import type { AlbumDetail } from "../../../../types/album-detail";
 import useLoadImage from "@/hooks/useLoadImage";
 import { formatArtists } from "@/lib/music/format-artists";
-import { toSongsWithAlbum } from "@/lib/mappers";
+import { toSongsWithAlbum } from "@/lib/mappers/song";
 import SongsGrid from "@/components/Song/SongsGrid";
 import { useUser } from "@/hooks/useUser";
 import renameAlbum from "@/actions/album/rename-album";
