+++ components/Modals/CreateAlbumModal.tsx
@@ -8,7 +8,7 @@ import { useUser } from "@/hooks/useUser";
 import type { AlbumWithArtists } from "../../types/album-with-artists";
 import type { Artist } from "../../types/artist";
 import { ALBUM_WITH_ARTISTS_SELECT } from "@/actions/_db-selects";
-import { mapAlbumWithArtistsRow } from "@/lib/mappers";
+import { mapAlbumWithArtistsRow } from "@/lib/mappers/album";
 import {
   Dialog,
   DialogContent,
